+++ platform/emulator/sit-check.cc
@@ -41,7 +41,7 @@ public:
   ScTrail() : FastStack() {}
   ~ScTrail() {}
   
-  void save(int * p) {
+  void save(uintptr_t * p) {
     // Save content and address
     push2((StackEntry) *p, (StackEntry) p);
   }
@@ -50,8 +50,8 @@ public:
     while (!isEmpty()) {
       StackEntry e1, e2;
       pop2(e1,e2);
-      int * p = (int *) e2;
-      int   v = (int)   e1;
+      uintptr_t * p = (uintptr_t *) e2;
+      uintptr_t   v = (uintptr_t)   e1;
       *p = v;
     } 
   }
@@ -87,10 +87,10 @@ static ScStack scStack;
  *
  */
 #define MARKVAR(u) \
-  scTrail.save((int *) u); *u=makeTaggedMarkInt(0);
+  scTrail.save((uintptr_t *) u); *u=makeTaggedMarkInt(0);
 
 #define MARKFIELD(d) \
-  scTrail.save((int *) d->cacGetMarkField()); d->cacMark(NULL);
+  scTrail.save((uintptr_t *) d->cacGetMarkField()); d->cacMark(NULL);
 
 /*
  * Test whether a space is good
